refactor longer example
