+++ markdown/testing/src/mockito/src/main/java/wuppie/mock/IWuppiWarenlager.java
@@ -1,4 +1,4 @@
-package stub;
+package wuppie.mock;
 
 import java.util.List;
