+++ markdown/testing/src/mockito/src/main/java/wuppie/mock/WuppiStore.java
@@ -1,7 +1,6 @@
-package spy;
+package wuppie.mock;
 
 import java.util.List;
-import mock.IWuppiWarenlager;
 
 /**
  * Über unsere brandheißen WuppiStores verkaufen wir dann alle Wuppis aus dem Wuppi-Warenlager (wenn
@@ -17,8 +16,8 @@ public class WuppiStore {
     /**
      * Bestellt alle Wuppis aus dem angegebenen Warenlager
      *
-     * @param lager Das lager aus dem die Wuppis geordert werden.
-     * @return Eine Liste mit allen Wuppis die bestellt wurden.
+     * @param lager Das Lager, aus dem die Wuppis geordert werden.
+     * @return Eine Liste mit allen Wuppis, die bestellt wurden.
      */
     public List<String> bestelleAlleWuppis(IWuppiWarenlager lager) {
         return lager.getAlleWuppis();
