+++ markdown/testing/src/mockito/src/main/java/wuppie/spy/WuppiStore.java
@@ -1,4 +1,4 @@
-package mock;
+package wuppie.spy;
 
 import java.util.List;
 
@@ -16,10 +16,10 @@ public class WuppiStore {
     /**
      * Bestellt alle Wuppis aus dem angegebenen Warenlager
      *
-     * @param lager Das Lager, aus dem die Wuppis geordert werden.
-     * @return Eine Liste mit allen Wuppis, die bestellt wurden.
+     * @param lager Das lager aus dem die Wuppis geordert werden.
+     * @return Eine Liste mit allen Wuppis die bestellt wurden.
      */
-    public List<String> bestelleAlleWuppis(IWuppiWarenlager lager) {
+    public List<Wuppi> bestelleAlleWuppis(IWuppiWarenlager lager) {
         return lager.getAlleWuppis();
     }
 }
