+++ markdown/testing/src/mockito/src/main/java/wuppie/stub/IWuppiWarenlager.java
@@ -1,4 +1,4 @@
-package mock;
+package wuppie.stub;
 
 import java.util.List;
