+++ markdown/testing/src/mockito/src/main/java/wuppie/stub/WuppiStore.java
@@ -1,4 +1,4 @@
-package stub;
+package wuppie.stub;
 
 import java.util.List;
