+++ markdown/testing/src/mockito/src/main/java/wuppie/verify/FluppisListe.java
@@ -1,4 +1,4 @@
-package verify;
+package wuppie.verify;
 
 import java.util.AbstractList;
